functions to deal mostly-atomically with Swift's Unmanaged
It is not possible to simply copy in Unmanaged pointer
and expect it to work if another thread could release
its last counted reference. It is necessary to have some
kind of lock in order to protect the incrementing of the
reference count in that case. Most other useful cases
involve swapping the pointer with an unrelated value, but
if we add the ability to lock for the "load" case we must
then account for locking in all other cases.